Extract disambiguation logic to function
